+++ Source/Core/Core/PowerPC/Jit64/Jit_FloatingPoint.cpp
@@ -13,6 +13,7 @@
 #include "Core/Core.h"
 #include "Core/PowerPC/Jit64/Jit.h"
 #include "Core/PowerPC/Jit64/JitRegCache.h"
+#include "Core/PowerPC/Jit64Common/Jit64PowerPCState.h"
 #include "Core/PowerPC/PPCAnalyst.h"
 #include "Core/PowerPC/PowerPC.h"
