+++ Source/Core/Core/PowerPC/Jit64Common/EmuCodeBlock.cpp
@@ -1,21 +1,44 @@
-// Copyright 2008 Dolphin Emulator Project
+// Copyright 2016 Dolphin Emulator Project
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/PowerPC/Jit64Common/Jit64Util.h"
-#include "Common/BitSet.h"
-#include "Common/CommonTypes.h"
+#include "Core/PowerPC/Jit64Common/EmuCodeBlock.h"
+
+#include "Common/Assert.h"
+#include "Common/CPUDetect.h"
 #include "Common/Intrinsics.h"
 #include "Common/MathUtil.h"
-#include "Common/x64ABI.h"
-#include "Common/x64Emitter.h"
 #include "Core/HW/MMIO.h"
 #include "Core/HW/Memmap.h"
+#include "Core/PowerPC/Gekko.h"
 #include "Core/PowerPC/Jit64Common/Jit64Base.h"
-#include "Core/PowerPC/Jit64Common/TrampolineCache.h"
+#include "Core/PowerPC/Jit64Common/Jit64PowerPCState.h"
 #include "Core/PowerPC/PowerPC.h"
 
 using namespace Gen;
+
+namespace
+{
+OpArg SwapImmediate(int access_size, const OpArg& reg_value)
+{
+  if (access_size == 32)
+    return Imm32(Common::swap32(reg_value.Imm32()));
+
+  if (access_size == 16)
+    return Imm16(Common::swap16(reg_value.Imm16()));
+
+  return Imm8(reg_value.Imm8());
+}
+
+OpArg FixImmediate(int access_size, OpArg arg)
+{
+  if (arg.IsImm())
+  {
+    arg = access_size == 8 ? arg.AsImm8() : access_size == 16 ? arg.AsImm16() : arg.AsImm32();
+  }
+  return arg;
+}
+}  // Anonymous namespace
 
 void EmuCodeBlock::MemoryExceptionCheck()
 {
@@ -47,6 +70,49 @@ void EmuCodeBlock::MemoryExceptionCheck()
   }
 }
 
+void EmuCodeBlock::SwitchToFarCode()
+{
+  nearcode = GetWritableCodePtr();
+  SetCodePtr(farcode.GetWritableCodePtr());
+}
+
+void EmuCodeBlock::SwitchToNearCode()
+{
+  farcode.SetCodePtr(GetWritableCodePtr());
+  SetCodePtr(nearcode);
+}
+
+FixupBranch EmuCodeBlock::CheckIfSafeAddress(const OpArg& reg_value, X64Reg reg_addr,
+                                             BitSet32 registers_in_use)
+{
+  registers_in_use[reg_addr] = true;
+  if (reg_value.IsSimpleReg())
+    registers_in_use[reg_value.GetSimpleReg()] = true;
+
+  // Get ourselves a free register; try to pick one that doesn't involve pushing, if we can.
+  X64Reg scratch = RSCRATCH;
+  if (!registers_in_use[RSCRATCH])
+    scratch = RSCRATCH;
+  else if (!registers_in_use[RSCRATCH_EXTRA])
+    scratch = RSCRATCH_EXTRA;
+  else
+    scratch = reg_addr;
+
+  if (scratch == reg_addr)
+    PUSH(scratch);
+  else
+    MOV(32, R(scratch), R(reg_addr));
+
+  // Perform lookup to see if we can use fast path.
+  SHR(32, R(scratch), Imm8(PowerPC::BAT_INDEX_SHIFT));
+  TEST(32, MScaled(scratch, SCALE_4, PtrOffset(&PowerPC::dbat_table[0])), Imm32(2));
+
+  if (scratch == reg_addr)
+    POP(scratch);
+
+  return J_CC(CC_Z, farcode.Enabled());
+}
+
 void EmuCodeBlock::UnsafeLoadRegToReg(X64Reg reg_addr, X64Reg reg_value, int accessSize, s32 offset,
                                       bool signExtend)
 {
@@ -61,6 +127,38 @@ void EmuCodeBlock::UnsafeLoadRegToRegNoSwap(X64Reg reg_addr, X64Reg reg_value, i
     MOVSX(32, accessSize, reg_value, MComplex(RMEM, reg_addr, SCALE_1, offset));
   else
     MOVZX(32, accessSize, reg_value, MComplex(RMEM, reg_addr, SCALE_1, offset));
+}
+
+void EmuCodeBlock::UnsafeWriteRegToReg(OpArg reg_value, X64Reg reg_addr, int accessSize, s32 offset,
+                                       bool swap, MovInfo* info)
+{
+  if (info)
+  {
+    info->address = GetWritableCodePtr();
+    info->nonAtomicSwapStore = false;
+  }
+
+  OpArg dest = MComplex(RMEM, reg_addr, SCALE_1, offset);
+  if (reg_value.IsImm())
+  {
+    if (swap)
+      reg_value = SwapImmediate(accessSize, reg_value);
+    MOV(accessSize, dest, reg_value);
+  }
+  else if (swap)
+  {
+    SwapAndStore(accessSize, dest, reg_value.GetSimpleReg(), info);
+  }
+  else
+  {
+    MOV(accessSize, dest, reg_value);
+  }
+}
+
+void EmuCodeBlock::UnsafeWriteRegToReg(Gen::X64Reg reg_value, Gen::X64Reg reg_addr, int accessSize,
+                                       s32 offset, bool swap, Gen::MovInfo* info)
+{
+  UnsafeWriteRegToReg(R(reg_value), reg_addr, accessSize, offset, swap, info);
 }
 
 bool EmuCodeBlock::UnsafeLoadToReg(X64Reg reg_value, OpArg opAddress, int accessSize, s32 offset,
@@ -102,6 +200,28 @@ bool EmuCodeBlock::UnsafeLoadToReg(X64Reg reg_value, OpArg opAddress, int access
 
   LoadAndSwap(accessSize, reg_value, memOperand, signExtend, info);
   return offsetAddedToAddress;
+}
+
+void EmuCodeBlock::UnsafeWriteGatherPipe(int accessSize)
+{
+  // No need to protect these, they don't touch any state
+  // question - should we inline them instead? Pro: Lose a CALL   Con: Code bloat
+  switch (accessSize)
+  {
+  case 8:
+    CALL(jit->GetAsmRoutines()->fifoDirectWrite8);
+    break;
+  case 16:
+    CALL(jit->GetAsmRoutines()->fifoDirectWrite16);
+    break;
+  case 32:
+    CALL(jit->GetAsmRoutines()->fifoDirectWrite32);
+    break;
+  case 64:
+    CALL(jit->GetAsmRoutines()->fifoDirectWrite64);
+    break;
+  }
+  jit->js.fifoBytesSinceCheck += accessSize >> 3;
 }
 
 // Visitor that generates code to read a MMIO value.
@@ -210,37 +330,6 @@ void EmuCodeBlock::MMIOLoadToReg(MMIO::Mapping* mmio, Gen::X64Reg reg_value,
     break;
   }
   }
-}
-
-FixupBranch EmuCodeBlock::CheckIfSafeAddress(const OpArg& reg_value, X64Reg reg_addr,
-                                             BitSet32 registers_in_use)
-{
-  registers_in_use[reg_addr] = true;
-  if (reg_value.IsSimpleReg())
-    registers_in_use[reg_value.GetSimpleReg()] = true;
-
-  // Get ourselves a free register; try to pick one that doesn't involve pushing, if we can.
-  X64Reg scratch = RSCRATCH;
-  if (!registers_in_use[RSCRATCH])
-    scratch = RSCRATCH;
-  else if (!registers_in_use[RSCRATCH_EXTRA])
-    scratch = RSCRATCH_EXTRA;
-  else
-    scratch = reg_addr;
-
-  if (scratch == reg_addr)
-    PUSH(scratch);
-  else
-    MOV(32, R(scratch), R(reg_addr));
-
-  // Perform lookup to see if we can use fast path.
-  SHR(32, R(scratch), Imm8(PowerPC::BAT_INDEX_SHIFT));
-  TEST(32, MScaled(scratch, SCALE_4, PtrOffset(&PowerPC::dbat_table[0])), Imm32(2));
-
-  if (scratch == reg_addr)
-    POP(scratch);
-
-  return J_CC(CC_Z, farcode.Enabled());
 }
 
 void EmuCodeBlock::SafeLoadToReg(X64Reg reg_value, const Gen::OpArg& opAddress, int accessSize,
@@ -399,119 +488,6 @@ void EmuCodeBlock::SafeLoadToRegImmediate(X64Reg reg_value, u32 address, int acc
   }
 }
 
-static OpArg SwapImmediate(int accessSize, const OpArg& reg_value)
-{
-  if (accessSize == 32)
-    return Imm32(Common::swap32(reg_value.Imm32()));
-  else if (accessSize == 16)
-    return Imm16(Common::swap16(reg_value.Imm16()));
-  else
-    return Imm8(reg_value.Imm8());
-}
-
-void EmuCodeBlock::UnsafeWriteRegToReg(OpArg reg_value, X64Reg reg_addr, int accessSize, s32 offset,
-                                       bool swap, MovInfo* info)
-{
-  if (info)
-  {
-    info->address = GetWritableCodePtr();
-    info->nonAtomicSwapStore = false;
-  }
-
-  OpArg dest = MComplex(RMEM, reg_addr, SCALE_1, offset);
-  if (reg_value.IsImm())
-  {
-    if (swap)
-      reg_value = SwapImmediate(accessSize, reg_value);
-    MOV(accessSize, dest, reg_value);
-  }
-  else if (swap)
-  {
-    SwapAndStore(accessSize, dest, reg_value.GetSimpleReg(), info);
-  }
-  else
-  {
-    MOV(accessSize, dest, reg_value);
-  }
-}
-
-static OpArg FixImmediate(int accessSize, OpArg arg)
-{
-  if (arg.IsImm())
-  {
-    arg = accessSize == 8 ? arg.AsImm8() : accessSize == 16 ? arg.AsImm16() : arg.AsImm32();
-  }
-  return arg;
-}
-
-void EmuCodeBlock::UnsafeWriteGatherPipe(int accessSize)
-{
-  // No need to protect these, they don't touch any state
-  // question - should we inline them instead? Pro: Lose a CALL   Con: Code bloat
-  switch (accessSize)
-  {
-  case 8:
-    CALL(jit->GetAsmRoutines()->fifoDirectWrite8);
-    break;
-  case 16:
-    CALL(jit->GetAsmRoutines()->fifoDirectWrite16);
-    break;
-  case 32:
-    CALL(jit->GetAsmRoutines()->fifoDirectWrite32);
-    break;
-  case 64:
-    CALL(jit->GetAsmRoutines()->fifoDirectWrite64);
-    break;
-  }
-  jit->js.fifoBytesSinceCheck += accessSize >> 3;
-}
-
-bool EmuCodeBlock::WriteToConstAddress(int accessSize, OpArg arg, u32 address,
-                                       BitSet32 registersInUse)
-{
-  arg = FixImmediate(accessSize, arg);
-
-  // If we already know the address through constant folding, we can do some
-  // fun tricks...
-  if (jit->jo.optimizeGatherPipe && PowerPC::IsOptimizableGatherPipeWrite(address))
-  {
-    if (!arg.IsSimpleReg(RSCRATCH))
-      MOV(accessSize, R(RSCRATCH), arg);
-
-    UnsafeWriteGatherPipe(accessSize);
-    return false;
-  }
-  else if (PowerPC::IsOptimizableRAMAddress(address))
-  {
-    WriteToConstRamAddress(accessSize, arg, address);
-    return false;
-  }
-  else
-  {
-    // Helps external systems know which instruction triggered the write
-    MOV(32, PPCSTATE(pc), Imm32(jit->js.compilerPC));
-
-    ABI_PushRegistersAndAdjustStack(registersInUse, 0);
-    switch (accessSize)
-    {
-    case 64:
-      ABI_CallFunctionAC(64, PowerPC::Write_U64, arg, address);
-      break;
-    case 32:
-      ABI_CallFunctionAC(32, PowerPC::Write_U32, arg, address);
-      break;
-    case 16:
-      ABI_CallFunctionAC(16, PowerPC::Write_U16, arg, address);
-      break;
-    case 8:
-      ABI_CallFunctionAC(8, PowerPC::Write_U8, arg, address);
-      break;
-    }
-    ABI_PopRegistersAndAdjustStack(registersInUse, 0);
-    return true;
-  }
-}
-
 void EmuCodeBlock::SafeWriteRegToReg(OpArg reg_value, X64Reg reg_addr, int accessSize, s32 offset,
                                      BitSet32 registersInUse, int flags)
 {
@@ -625,6 +601,63 @@ void EmuCodeBlock::SafeWriteRegToReg(OpArg reg_value, X64Reg reg_addr, int acces
   }
 }
 
+void EmuCodeBlock::SafeWriteRegToReg(Gen::X64Reg reg_value, Gen::X64Reg reg_addr, int accessSize,
+                                     s32 offset, BitSet32 registersInUse, int flags)
+{
+  SafeWriteRegToReg(R(reg_value), reg_addr, accessSize, offset, registersInUse, flags);
+}
+
+bool EmuCodeBlock::WriteClobbersRegValue(int accessSize, bool swap)
+{
+  return swap && !cpu_info.bMOVBE && accessSize > 8;
+}
+
+bool EmuCodeBlock::WriteToConstAddress(int accessSize, OpArg arg, u32 address,
+                                       BitSet32 registersInUse)
+{
+  arg = FixImmediate(accessSize, arg);
+
+  // If we already know the address through constant folding, we can do some
+  // fun tricks...
+  if (jit->jo.optimizeGatherPipe && PowerPC::IsOptimizableGatherPipeWrite(address))
+  {
+    if (!arg.IsSimpleReg(RSCRATCH))
+      MOV(accessSize, R(RSCRATCH), arg);
+
+    UnsafeWriteGatherPipe(accessSize);
+    return false;
+  }
+  else if (PowerPC::IsOptimizableRAMAddress(address))
+  {
+    WriteToConstRamAddress(accessSize, arg, address);
+    return false;
+  }
+  else
+  {
+    // Helps external systems know which instruction triggered the write
+    MOV(32, PPCSTATE(pc), Imm32(jit->js.compilerPC));
+
+    ABI_PushRegistersAndAdjustStack(registersInUse, 0);
+    switch (accessSize)
+    {
+    case 64:
+      ABI_CallFunctionAC(64, PowerPC::Write_U64, arg, address);
+      break;
+    case 32:
+      ABI_CallFunctionAC(32, PowerPC::Write_U32, arg, address);
+      break;
+    case 16:
+      ABI_CallFunctionAC(16, PowerPC::Write_U16, arg, address);
+      break;
+    case 8:
+      ABI_CallFunctionAC(8, PowerPC::Write_U8, arg, address);
+      break;
+    }
+    ABI_PopRegistersAndAdjustStack(registersInUse, 0);
+    return true;
+  }
+}
+
 void EmuCodeBlock::WriteToConstRamAddress(int accessSize, OpArg arg, u32 address, bool swap)
 {
   X64Reg reg;
@@ -651,6 +684,30 @@ void EmuCodeBlock::WriteToConstRamAddress(int accessSize, OpArg arg, u32 address
     SwapAndStore(accessSize, MRegSum(RMEM, RSCRATCH2), reg);
   else
     MOV(accessSize, MRegSum(RMEM, RSCRATCH2), R(reg));
+}
+
+void EmuCodeBlock::JitGetAndClearCAOV(bool oe)
+{
+  if (oe)
+    AND(8, PPCSTATE(xer_so_ov), Imm8(~XER_OV_MASK));  // XER.OV = 0
+  SHR(8, PPCSTATE(xer_ca), Imm8(1));                  // carry = XER.CA, XER.CA = 0
+}
+
+void EmuCodeBlock::JitSetCA()
+{
+  MOV(8, PPCSTATE(xer_ca), Imm8(1));  // XER.CA = 1
+}
+
+// Some testing shows CA is set roughly ~1/3 of the time (relative to clears), so
+// branchless calculation of CA is probably faster in general.
+void EmuCodeBlock::JitSetCAIf(CCFlags conditionCode)
+{
+  SETcc(conditionCode, PPCSTATE(xer_ca));
+}
+
+void EmuCodeBlock::JitClearCA()
+{
+  MOV(8, PPCSTATE(xer_ca), Imm8(0));
 }
 
 void EmuCodeBlock::ForceSinglePrecision(X64Reg output, const OpArg& input, bool packed,
@@ -1081,30 +1138,6 @@ void EmuCodeBlock::SetFPRF(Gen::X64Reg xmm)
   SetJumpTarget(continue4);
   SHL(32, R(RSCRATCH), Imm8(FPRF_SHIFT));
   OR(32, PPCSTATE(fpscr), R(RSCRATCH));
-}
-
-void EmuCodeBlock::JitGetAndClearCAOV(bool oe)
-{
-  if (oe)
-    AND(8, PPCSTATE(xer_so_ov), Imm8(~XER_OV_MASK));  // XER.OV = 0
-  SHR(8, PPCSTATE(xer_ca), Imm8(1));                  // carry = XER.CA, XER.CA = 0
-}
-
-void EmuCodeBlock::JitSetCA()
-{
-  MOV(8, PPCSTATE(xer_ca), Imm8(1));  // XER.CA = 1
-}
-
-// Some testing shows CA is set roughly ~1/3 of the time (relative to clears), so
-// branchless calculation of CA is probably faster in general.
-void EmuCodeBlock::JitSetCAIf(CCFlags conditionCode)
-{
-  SETcc(conditionCode, PPCSTATE(xer_ca));
-}
-
-void EmuCodeBlock::JitClearCA()
-{
-  MOV(8, PPCSTATE(xer_ca), Imm8(0));
 }
 
 void EmuCodeBlock::Clear()
