+++ Source/Core/Core/PowerPC/Jit64Common/TrampolineCache.h
@@ -5,7 +5,14 @@
 #pragma once
 
 #include "Common/CommonTypes.h"
-#include "Core/PowerPC/Jit64Common/Jit64Util.h"
+#include "Core/PowerPC/Jit64Common/EmuCodeBlock.h"
+
+struct TrampolineInfo;
+
+// a bit of a hack; the MMU results in more code ending up in the trampoline cache,
+// because fastmem results in far more backpatches in MMU mode
+constexpr int TRAMPOLINE_CODE_SIZE = 1024 * 1024 * 8;
+constexpr int TRAMPOLINE_CODE_SIZE_MMU = 1024 * 1024 * 32;
 
 // We need at least this many bytes for backpatching.
 constexpr int BACKPATCH_SIZE = 5;
